requirements.txt
ccxt
pandas
numpy
yfinance
schedule
python-dotenv
termcolor
requests
